// src/client/components/pages/home/HeroHome.module.scss.d.ts
/** @format */

// Auto generated by typed-scss-modules
export type Styles = {
    'color-black': string;
    'color-primary': string;
    'color-secondary': string;
    'color-white': string;
    gradient: string;
    hero: string;
    hero__bg: string;
    hero__bg_outline: string;
    hero__btn: string;
    hero__heading: string;
    hero__img: string;
    hero__img_container: string;
    hero__info_container: string;
    hero__intro: string;
    wave: string;
};

export type ClassNames = keyof Styles;

declare const styles: Styles;

export default styles;